backend/requirements.txt
genlayer-py==0.8.1
httpx==0.27.0
python-dotenv==1.0.1
schedule==1.2.1

